Add module sale_procurement_group_by_line
This module extract logic to create multiple procurement group
for a single sale order for grouped sale order lines.
