+++ sale_procurement_group_by_line/README.rst
@@ -1,0 +1,54 @@
+.. image:: https://img.shields.io/badge/licence-AGPL--3-blue.svg
+   :alt: License: AGPL-3
+
+Sale Procurement Group by line
+==============================
+
+This module was written to extend the functionality of procurement groups
+created from a Sale order.
+
+On itself, this module does nothing it is a requirement for modules which
+needs to create procurement group per sale order line basis.
+
+Installation
+============
+
+This module is only a base for other modules, see Usage.
+
+
+Configuration
+=============
+
+No special configuration is required.
+
+Usage
+=====
+
+To use this module, you need to install
+**Sale Sourcing By Line** (`sale_sourcing_by_line`) and/or
+**Sale Quotation Sourcing** (`sale_quotation_sourcing`)
+
+
+Credits
+=======
+
+Contributors
+------------
+
+* Yannick Vaucher <yannick.vaucher@camptocamp.com>
+* Guewen Baconnier <guewen.baconnier@camptocamp.com>
+
+Maintainer
+----------
+
+.. image:: http://odoo-community.org/logo.png
+   :alt: Odoo Community Association
+   :target: http://odoo-community.org
+
+This module is maintained by the OCA.
+
+OCA, or the Odoo Community Association, is a nonprofit organization whose
+mission is to support the collaborative development of Odoo features and
+promote its widespread use.
+
+To contribute to this module, please visit http://odoo-community.org.
