+++ sale_procurement_group_by_line/model/__init__.py
@@ -1,0 +1,1 @@
+from . import sale
