+++ sale_sourced_by_line/README.rst
@@ -1,0 +1,39 @@
+.. image:: https://img.shields.io/badge/licence-AGPL--3-blue.svg
+   :alt: License: AGPL-3
+
+Sale Sourced by Line
+====================
+
+This module adds the possibility to source a line of sale order from a specific
+warehouse instead of using the warehouse of the sale order.
+
+This will create one procurement group per warehouse set in sale
+order lines.
+
+It will only supports routes such as MTO and Drop Shipping.
+
+Credits
+=======
+
+Contributors
+------------
+
+* Guewen Baconnier <guewen.baconnier@camptocamp.com>
+* Yannick Vaucher <yannick.vaucher@camptocamp.com>
+
+Maintainer
+----------
+
+.. image:: http://odoo-community.org/logo.png
+   :alt: Odoo Community Association
+   :target: http://odoo-community.org
+
+This module is maintained by the OCA.
+
+OCA, or the Odoo Community Association, is a nonprofit organization whose
+mission is to support the collaborative development of Odoo features and
+promote its widespread use.
+
+To contribute to this module, please visit http://odoo-community.org.
+
+
